src/requirements.txt
# Core ML
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Data
yfinance>=0.2.40
pandas>=2.0.0

# API
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# Serialization
joblib>=1.3.0
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0

# Dev
pytest>=7.0.0
